+++ tasmania/python/grids/grid.py
@@ -34,7 +34,7 @@ from tasmania.python.utils.utils import smaller_than as lt, smaller_or_equal_tha
 try:
     from tasmania.conf import datatype
 except ImportError:
-    datatype = np.float32
+    datatype = np.float64
 
 
 class Grid:
