+++ tasmania/python/isentropic/physics/vertical_advection.py
@@ -38,7 +38,7 @@ from tasmania.python.utils.storage_utils import zeros
 try:
     from tasmania.conf import datatype
 except ImportError:
-    datatype = np.float32
+    datatype = np.float64
 
 
 mfwv = "mass_fraction_of_water_vapor_in_air"
